requirements.txt
streamlit
crewai
crewai-tools
langfuse
openai
requests
python-dotenv
beautifulsoup4
langfuse<=2.25.0